requirements.txt
pillow>=10.0.0
reportlab>=4.0.0